pyproject.toml
[project]
name = "kian"
version = "0.1.0"
description = "Voice assistant: VAD → Whisper → Qwen3.5-4B → Piper TTS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "webrtcvad>=2.0.10",
    "faster-whisper>=1.1.0",
    "llama-cpp-python>=0.3.19",
    "piper-tts>=1.4",
    "sounddevice>=0.5",
    "numpy>=1.24",
    "huggingface-hub>=1.8.0",
    "setuptools<82",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
kian = "kian.app:main"

# [[tool.uv.index]]
# url = "https://pypi.jetson-ai-lab.io/jp6/cu126"
